Add skeleton front-end app
The client app is currently just a login page that does nothing. It is
rendered with React and packaged via webpack.

Babel configuration has been refactored out into `.babelrc.js` so that
webpack (via `webpack.config.js`) can use this configuration while
building the app.

New gulp tasks have been added to build the JS for the front-end. The
server now reloads the front-end when it is changed.