Migrate to gulp4 and add PID logging
The gulpfile has been migrated to gulp4. In addition, the server now
keeps track of its pid in a file (`.server.pid`). If this file exists
when `gulp serve` is run, we will attempt to kill the process outlined
in that file.

All operations that update the webpack configuration now take place in
`webpack.config.js`, so running `gulp build --production` will now have
the same result as running `webpack` when `NODE_ENV` is set to
`production`.