Add a utility to determine if an error is a unique constraint error
sqlite and PostgreSQL have different error codes for reporting if an
error is a violation of a `UNIQUE` constraint. The check for this has
been refactored into a utility that checks for the PostgreSQL error code
when `NODE_ENV` is set to production and the sqlite error code
otherwise.

In addition, the `User.create` method will now return a `Promise` that
resolves to `null` if it encounters a `UNIQUE` constraint when creating
a user; otherwise it will return the `id` of the user.