Handle error correctly in the server app
Previously, we were not logging errors when we caught them. We now log
them with the error level. In addition, we no longer emit the `error`
event as that results in the error being printed to the console twice.
When we handle an error this way, we return a JSON response to the
client.