Simplify the MessageList behaviour with unique IDs
Instead of providing an addition `uniqueID` parameter to the
`addMessage` action, it now takes the `id` parameter to optionally set
the `id` of the message to a string. If not provided (or a non-string
value is provided), an incrementing integer value will be used instead.

Because of this simplification, the `dismissMessageByUniqueId` action
has been removed as it is no longer necessary.