Add a <Label> component; use it in <LoginForm> and <RegisterForm>
The `<Label>` component shows a floating label in front of the text
input. When the input has no text, the label will be the size of the
text and centred vertically. When the field *does* have text, the label
will shrink (via an animation) to the upper left so that text can be entered.

The `<Label>` component uses an actual `<label>` underneath to improve
accessibility with screen readers.